requirements.txt
Django>=5.0
qrcode
djangorestframework>=3.15
drf-yasg>=1.21.0
requests>=2.28
# Dev / testing
pytest>=7.0
pytest-django>=4.5
